requirements.txt
black==22.8.0
cryptography==36.0.1
dash==2.9.3
dash-bootstrap-components==1.1.0
dash-deck==0.0.1
dash-extensions==1.0.1
dash-iconify==0.1.2
dash-mantine-components==0.12.1
Flask==2.1.1
gunicorn==20.1.0
ipykernel==6.19.2
itsdangerous==2.1.2
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
plotly==5.7.0
pydeck==0.8.0
python-dotenv==0.13.0
werkzeug==2.2.2